requirements.txt
transformers
torch
PyPDF2
python-docx
tqdm
pymupdf
youtube-transcript-api
pytube